Renamed "PackingListPreset" to "Preset" (collection and api including all dependencies)
